+++ hoot-core/src/main/cpp/hoot/core/scoring/BaseComparator.cpp
@@ -69,17 +69,17 @@ void BaseComparator::_calculateColor(double v, double max, QRgb& c) const
     if (v < max / 3.0)
     {
       double r = v / (max / 3.0) * 255.0;
-      c = qRgb(r, 0, 0);
+      c = qRgb((int)r, 0, 0);
     }
     else if (v < max * 2.0 / 3.0)
     {
       double r = (v - (max / 3.0)) / (max / 3.0) * 255.0;
-      c = qRgb(255, r, 0);
+      c = qRgb(255, (int)r, 0);
     }
     else if (v <= max)
     {
       double r = (v - (max * 2.0 / 3.0)) / (max / 3.0) * 255.0;
-      c = qRgb(255, 255, r);
+      c = qRgb(255, 255, (int)r);
     }
     else
     {
